requirements.txt
paramiko==3.5.0
PyYAML==6.0.2
rich==13.9.4
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
pysnmp==4.4.12
psutil==5.9.6
prometheus-client==0.19.0
cryptography==41.0.7
